-Add: ZPool properties.
Note, this requires a (valid) active handle to libzfs, hence the
decorator.
The difference over other libs (like pyzfscore) is that property output
can be up to MAXNAMELEN data.. if we do not allocate enough data, we
might just only get the first bit of data (8 bytes, in most cases)...
this might work for non-literal values, but not for literal.